LPD-19855 refactor to remove helper
